renamed WebWorkerBehavior to UserWorkerBehaivor and moved to workers.rs
